Move port in use exception test as it does not apply to Tomcat
See gh-8737
